+++ packages/core/src/plugins.ts
@@ -380,7 +380,7 @@ export function makeInlinePluginsCreator(
         debug('all released, push changed file to git')
         const makePushToGit = plugins.makePushToGit
         // eslint-disable-next-line no-console
-        console.log(context.branch, makePushToGit)
+        console.log(makePushToGit)
         if (makePushToGit) {
           const releases = pkgs.map((pkg) => ({
             lastRelease: lastReleaseMap[pkg],
@@ -388,7 +388,7 @@ export function makeInlinePluginsCreator(
           }))
           const pushToGit = await makePushToGit(context.branch, releases)
           // eslint-disable-next-line no-console
-          console.log(context.branch, releases, pushToGit)
+          console.log(releases, pushToGit)
           await pushToGit({ ...context, cwd: process.cwd() })
         }