+++ packages/core/CHANGELOG.md
@@ -1,3 +1,5 @@
+## @semantic-release-monorepo/core [2.1.7](https://github.com/bubkoo/semantic-release-monorepo/compare/@semantic-release-monorepo/core@2.1.6...@semantic-release-monorepo/core@2.1.7) (2022-11-25)
+
 ## @semantic-release-monorepo/core [2.1.6](https://github.com/bubkoo/semantic-release-monorepo/compare/@semantic-release-monorepo/core@2.1.5...@semantic-release-monorepo/core@2.1.6) (2022-11-25)
 
 ## @semantic-release-monorepo/core [2.1.5](https://github.com/bubkoo/semantic-release-monorepo/compare/@semantic-release-monorepo/core@2.1.4...@semantic-release-monorepo/core@2.1.5) (2022-11-25)